+++ test/hover/init.lua
@@ -890,6 +890,21 @@ local t: {
 }
 ]]
 
+TEST [[
+local <?t?> = {
+    [-1] = -1,
+    [0]  = 0,
+    [1]  = 1,
+}
+]]
+[[
+local t: {
+    [-1]: integer = -1,
+    [0]: integer = 0,
+    [1]: integer = 1,
+}
+]]
+
 TEST[[
 ---@class Class
 local <?x?> = class()