+++ changelog.md
@@ -5,6 +5,7 @@
 * `NEW` fully supports `---@generic T`
 * `FIX` [#274](https://github.com/sumneko/lua-language-server/issues/274)
 * `FIX` [#279](https://github.com/sumneko/lua-language-server/issues/279)
+* `FIX` [#280](https://github.com/sumneko/lua-language-server/issues/280)
 
 ## 1.4.0
 `2020-12-3`